Upcast AbstractProject in private functions
Change the private methods that accept an AbstractBuild parameter so they
accept BuildableItem instead, upcasting the object so these methods may be
reused more readily.

While we could upcast all the way to SubTask, the interface that actually
specifies getLastBuiltOn(), that would not add much value, as this class
does not use SubTask.  Indeed, we would need to import SubTask just to use
it in this fashion.  Conversely, BuildableItem is used already and forms
the base of this Trigger<> object.