Remove unnecessary downcasting
The job object is cast from a BuildableItem into an AbstractProject, then
that AbstractProject calls getAssignedLabel().  However, the function is
specified in the BuildableItem interface; therefore, the original object
is sufficient without casting down to AbstractProject.  By removing this
unnecessary downcast, the program has one less runtime check to make.