Transition to Jenkins singleton, from Hudson
In Jenkins, the Hudson object is a subclass of Jenkins, one that provides
only deprecated functions (presumably for backwards compatibility).  Since
xtrigger-lib does not use the deprecated functions, transition the project
over to using the Jenkins object.

While this necessitates a version boost in the requirements, it is a small
one.  And it allows the project to move forward with a more modern (well,
less antiquated) implementation.

Note: In Jenkins 2.98, a new method, get(), replaces getActiveInstance()
and the getActiveInstance() method is now deprecated.  In future, another
jump forward will likely be necessary.